+++ docs/src/components/common/Footer/index.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { FiGithub } from 'react-icons/fi';
 
 import styles from './styles.module.css';