+++ docs/src/components/forward/Link.tsx
@@ -1,5 +1,5 @@
 import NextLink, { LinkProps } from 'next/link';
-import React, { HTMLAttributes } from 'react';
+import { HTMLAttributes } from 'react';
 import { UrlObject } from 'url';
 
 interface Props extends LinkProps, HTMLAttributes<HTMLElement> {}