+++ docs/src/pages/_app.tsx
@@ -1,6 +1,5 @@
 import { MDXProvider } from '@mdx-js/react';
 import App from 'next/app';
-import React from 'react';
 
 import { Head } from '@components/common/Head';
 import { Layout } from '@components/layout/Layout';