+++ reactools/__tests__/resources/mergeRefs.spec.ts
@@ -8,7 +8,7 @@ const data = {} as const;
 type Data = typeof data;
 
 describe('mergeRefs', () => {
-  it('should return the value passed to the hook', () => {
+  it('should receive the value passed to the hook', () => {
     const { result: ref } = renderHook(() => useRef<Data>(null));
 
     act(() => {
@@ -18,7 +18,7 @@ describe('mergeRefs', () => {
     expect(ref.current.current).toEqual(data);
   });
 
-  it('should return the value passed to the hook in all references passed', () => {
+  it('should receive the value passed to the hook in all references passed', () => {
     const { result: ref1 } = renderHook(() => useRef<Data>(null));
     const { result: ref2 } = renderHook(() => useRef<Data>(null));
 
@@ -31,7 +31,7 @@ describe('mergeRefs', () => {
     expect(ref1.current.current).toEqual(ref2.current.current);
   });
 
-  it('should return the value passed to the hook using function reference', () => {
+  it('should receive the value passed to the hook using function reference', () => {
     const { result: ref } = renderHook(
       () => createRef<Data>() as MutableRefObject<Data>,
     );
@@ -46,7 +46,7 @@ describe('mergeRefs', () => {
     expect(ref.current.current).toEqual(data);
   });
 
-  it('should not return the value because it is an invalid reference', () => {
+  it('should not receive the value because it is an invalid reference', () => {
     const ref = Symbol('ref');
 
     act(() => {
@@ -56,7 +56,7 @@ describe('mergeRefs', () => {
     expect((ref as any).current).not.toEqual(data);
   });
 
-  it('should return the value passed to the hook because the object is editable', () => {
+  it('should receive the value passed to the hook because the object is editable', () => {
     const ref = {} as MutableRefObject<Data>;
 
     act(() => {